fix: gate wear debug clock to debug builds and tick today screen
- Guard WearApplication and DebugClockListener with BuildConfig.DEBUG so
  a stale /tigerduck/debug-clock entry can't apply fake time on a
  release watch (mirrors DebugClockController.bootstrap).
- Enable buildConfig in :wear so BuildConfig.DEBUG is generated.
- Share currentTaipeiTick between NowNextScreen and TodayScreen via a
  new TaipeiTick.kt helper so class status badges refresh on minute
  boundaries without the user navigating away.
- Replace single-slot onAccentColorChanged callback with a SharedFlow
  to match the other change-flows in AppPreferences and avoid silently
  evicting the wear-bridge subscriber if another caller registers.
