+++ wear/src/main/java/org/ntust/app/tigerduck/wear/ui/TodayScreen.kt
@@ -31,7 +31,6 @@ import org.ntust.app.tigerduck.shared.TodayClassEntry
 import org.ntust.app.tigerduck.shared.TodayClassStatus
 import org.ntust.app.tigerduck.wear.R
 import org.ntust.app.tigerduck.wear.data.WatchSnapshot
-import org.ntust.app.tigerduck.shared.clock.AppClock
 import org.ntust.app.tigerduck.wear.ui.theme.LocalAccentColor
 import org.ntust.app.tigerduck.wear.ui.theme.LocalScreenPadding
 
@@ -41,9 +40,9 @@ fun TodayScreen(
     onRowClick: (String) -> Unit,
 ) {
     val pad = LocalScreenPadding.current
-    val now = AppClock.localDateTime()
-    val weekday = now.dayOfWeek.value
-    val minuteOfDay = now.hour * 60 + now.minute
+    // Tick once a minute so a class transitioning to Ended (and losing its
+    // Ongoing highlight) refreshes without the user navigating away.
+    val (weekday, minuteOfDay) = currentTaipeiTick()
     val entries = NextClassResolver.todaysClasses(snapshot.courses, weekday, minuteOfDay)
     val listState = rememberScalingLazyListState()
